Module ethers_providers::ens

Ethereum Name Service support Adapted from https://github.com/hhatto/rust-ens/blob/master/src/lib.rs

Constants

ADDR_SELECTOR

addr(bytes32)

ENS_ADDRESS

ENS registry address (0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e)

FIELD_SELECTOR

text(bytes32, string)

INTERFACE_SELECTOR

supportsInterface(bytes4 interfaceID)

NAME_SELECTOR

name(bytes32)

Functions

bytes_32ify

Returns a number in bytes form with padding to fit in 32 bytes.

get_resolver

Returns a transaction request for calling the resolver method on the ENS server

namehash

Returns the ENS namehash as specified in EIP-137

parameterhash

Returns the ENS record key hash EIP-634

resolve

Returns a transaction request for calling

reverse_address

Returns the reverse-registrar name of an address.

supports_interface

Returns a transaction request for checking interface support