Struct ethers_providers::PendingTransaction

pub struct PendingTransaction<'a, P> { /* fields omitted */ }

A pending transaction is a transaction which has been submitted but is not yet mined. await’ing on a pending transaction will resolve to a transaction receipt once the transaction has enough confirmations. The default number of confirmations is 1, but may be adjusted with the confirmations method. If the transaction does not have enough confirmations or is not mined, the future will stay in the pending state.

Implementations

impl<'a, P: JsonRpcClient> PendingTransaction<'a, P>

pub fn new(tx_hash: TxHash, provider: &'a Provider<P>) -> Self

Creates a new pending transaction poller from a hash and a provider

pub fn provider(&self) -> Provider<P> where
P: Clone,

Returns the Provider associated with the pending transaction

pub fn confirmations(self, confs: usize) -> Self

Sets the number of confirmations for the pending transaction to resolve to a receipt

pub fn interval<T: Into<Duration>>(self, duration: T) -> Self

Sets the polling interval