Struct ethers_providers::Provider

pub struct Provider<P> { /* private fields */ }

An abstract provider for interacting with the Ethereum JSON RPC API. Must be instantiated with a data transport which implements the JsonRpcClient trait (e.g. HTTP, Websockets etc.)

Example

use ethers_providers::{Middleware, Provider, Http};
use std::convert::TryFrom;

let provider = Provider::<Http>::try_from(
    "https://eth.llamarpc.com"
).expect("could not instantiate HTTP Provider");

let block = provider.get_block(100u64).await?;
println!("Got block: {}", serde_json::to_string(&block)?);

Implementations

impl<P: JsonRpcClient> Provider<P>

pub fn new(provider: P) -> Self

Instantiate a new provider with a backend.

pub async fn node_client(&self) -> Result<NodeClient, ProviderError>

Returns the type of node we’re connected to, while also caching the value for use in other node-specific API calls, such as the get_block_receipts call.

pub fn with_sender(self, address: impl Into<Address>) -> Self

Set the default sender on the provider

pub async fn request<T, R>( &self, method: &str, params: T ) -> Result<R, ProviderError>
where T: Debug + Serialize + Send + Sync, R: Serialize + DeserializeOwned + Debug + Send,

Make an RPC request via the internal connection, and return the result.

pub fn call_raw<'a>(&'a self, tx: &'a TypedTransaction) -> CallBuilder<'a, P>

Analogous to Middleware::call, but returns a CallBuilder that can either be .awaitd or used to override the parameters sent to eth_call.

See the ethers_core::types::spoof for functions to construct state override parameters.

Note: this method does not send a transaction from your account

Example

let geth = Geth::new().spawn();
let provider = Provider::<Http>::try_from(geth.endpoint()).unwrap();

let adr1: Address = "0x6fC21092DA55B392b045eD78F4732bff3C580e2c".parse()?;
let adr2: Address = "0x295a70b2de5e3953354a6a8344e616ed314d7251".parse()?;
let pay_amt = parse_ether(1u64)?;

// Not enough ether to pay for the transaction
let tx = TransactionRequest::pay(adr2, pay_amt).from(adr1).into();

// override the sender's balance for the call
let mut state = spoof::balance(adr1, pay_amt * 2);
provider.call_raw(&tx).state(&state).await?;

impl<P: JsonRpcClient> Provider<P>

pub fn ens<T: Into<Address>>(self, ens: T) -> Self

Sets the ENS Address (default: mainnet)

pub fn set_interval<T: Into<Duration>>(&mut self, interval: T) -> &mut Self

Sets the default polling interval for event filters and pending transactions (default: 7 seconds)

pub fn interval<T: Into<Duration>>(self, interval: T) -> Self

Sets the default polling interval for event filters and pending transactions (default: 7 seconds)

pub fn get_interval(&self) -> Duration

Gets the polling interval which the provider currently uses for event filters and pending transactions (default: 7 seconds)

impl Provider<Http>

pub fn url(&self) -> &Url

The Url to which requests are made

pub fn url_mut(&mut self) -> &mut Url

Mutable access to the Url to which requests are made

impl<Read, Write> Provider<RwClient<Read, Write>>
where Read: JsonRpcClient + 'static, <Read as JsonRpcClient>::Error: Sync + Send + 'static, Write: JsonRpcClient + 'static, <Write as JsonRpcClient>::Error: Sync + Send + 'static,

pub fn rw(r: Read, w: Write) -> Self

Creates a new Provider with a RwClient

impl<T: JsonRpcClientWrapper> Provider<QuorumProvider<T>>

pub fn quorum(inner: QuorumProvider<T>) -> Self

Provider that uses a quorum

impl Provider<MockProvider>

pub fn mocked() -> (Self, MockProvider)

Returns a Provider instantiated with an internal “mock” transport.

Example

use ethers_core::types::U64;
use ethers_providers::{Middleware, Provider};
// Instantiate the provider
let (provider, mock) = Provider::mocked();
// Push the mock response
mock.push(U64::from(12))?;
// Make the call
let blk = provider.get_block_number().await.unwrap();
// The response matches
assert_eq!(blk.as_u64(), 12);
// and the request as well!
mock.assert_request("eth_blockNumber", ()).unwrap();

impl Provider<RetryClient<Http>>

pub fn new_client( src: &str, max_retry: u32, initial_backoff: u64 ) -> Result<Self, ParseError>

Create a new RetryClient by connecting to the provided URL. Errors if src is not a valid URL