Struct ethers_providers::RetryClient

pub struct RetryClient<T>where
    T: JsonRpcClient,
    T::Error: RpcError + Sync + Send + 'static,{ /* private fields */ }

RetryClient presents as a wrapper around JsonRpcClient that will retry requests based with an exponential backoff and filtering based on RetryPolicy.

The RetryPolicy, mainly for rate-limiting errors, can be adjusted for specific applications, endpoints. In addition to the RetryPolicy errors due to connectivity issues, like timed out connections or responses in range 5xx can be retried separately.

Example

use ethers_providers::{Http, RetryClient, RetryClientBuilder, HttpRateLimitRetryPolicy};
use std::time::Duration;
use url::Url;

let http = Http::new(Url::parse("http://localhost:8545").unwrap());
let client = RetryClientBuilder::default()
    .rate_limit_retries(10)
    .timeout_retries(3)
    .initial_backoff(Duration::from_millis(500))
    .build(http, Box::new(HttpRateLimitRetryPolicy::default()));

Implementations

impl<T> RetryClient<T>where T: JsonRpcClient, T::Error: Sync + Send + 'static,

pub fn new( inner: T, policy: Box<dyn RetryPolicy<T::Error>>, max_retry: u32, initial_backoff: u64 ) -> Self

Creates a new RetryClient that wraps a client and adds retry and backoff support

Example

 
use ethers_providers::{Http, RetryClient, HttpRateLimitRetryPolicy};
use std::time::Duration;
use url::Url;

let http = Http::new(Url::parse("http://localhost:8545").unwrap());
let backoff_timeout = 3000; // in ms
let max_retries = 10;
let client = RetryClient::new(http, Box::new(HttpRateLimitRetryPolicy::default()), max_retries, backoff_timeout);

pub fn set_compute_units(&mut self, cpus: u64) -> &mut Self

Sets the free compute units per second limit.

This is the maximum number of weighted request that can be handled per second by the endpoint before rate limit kicks in.

This is used to guesstimate how long to wait until to retry again