Struct ethers_providers::Provider

pub struct Provider<P> { /* fields omitted */ }

An abstract provider for interacting with the Ethereum JSON RPC API. Must be instantiated with a data transport which implements the JsonRpcClient trait (e.g. HTTP, Websockets etc.)

Example

use ethers_providers::{Middleware, Provider, Http};
use std::convert::TryFrom;

let provider = Provider::<Http>::try_from(
    "https://mainnet.infura.io/v3/c60b0bb42f8a4c6481ecd229eddaca27"
).expect("could not instantiate HTTP Provider");

let block = provider.get_block(100u64).await?;
println!("Got block: {}", serde_json::to_string(&block)?);

Implementations

impl<P: JsonRpcClient> Provider<P>

pub fn new(provider: P) -> Self

Instantiate a new provider with a backend.

pub async fn node_client(&self) -> Result<NodeClient, ProviderError>

Returns the type of node we’re connected to, while also caching the value for use in other node-specific API calls, such as the get_block_receipts call.

pub fn with_sender(self, address: impl Into<Address>) -> Self

impl<P: JsonRpcClient> Provider<P>

pub fn ens<T: Into<Address>>(self, ens: T) -> Self

Sets the ENS Address (default: mainnet)

pub fn interval<T: Into<Duration>>(self, interval: T) -> Self

Sets the default polling interval for event filters and pending transactions (default: 7 seconds)

pub fn get_interval(&self) -> Duration

Gets the polling interval which the provider currently uses for event filters and pending transactions (default: 7 seconds)

impl<T: JsonRpcClientWrapper> Provider<QuorumProvider<T>>

pub fn quorum(inner: QuorumProvider<T>) -> Self

Provider that uses a quorum

impl Provider<MockProvider>

pub fn mocked() -> (Self, MockProvider)

Returns a Provider instantiated with an internal “mock” transport.

Example

use ethers_core::types::U64;
use ethers_providers::{Middleware, Provider};
// Instantiate the provider
let (provider, mock) = Provider::mocked();
// Push the mock response
mock.push(U64::from(12))?;
// Make the call
let blk = provider.get_block_number().await.unwrap();
// The response matches
assert_eq!(blk.as_u64(), 12);
// and the request as well!
mock.assert_request("eth_blockNumber", ()).unwrap();